removes unused function concatenates conditionals and moves function to same file is is used in
